temp fix for 2048
i hate doing these temp fixes, but it works
in case anyone acc gives a damn abt these descriptions, whats happening is that the index html in the game file is doing /folder/file when in reality i need folder/file. idk where to find the /folder/file to fix the problem. so instead, i import the files into replit as a hosting thing for a temp fix. its not the best, but it'll work for now.